+++ lib/src/models/options.dart
@@ -42,6 +42,12 @@ class MapBoxOptions {
   /// The unit of measure said in voice instructions
   final VoiceUnits? units;
 
+  final double? maxHeight;
+
+  final double? maxWeight;
+
+  final double? maxLength;
+
   /// If the value of this property is true, a returned route may require an immediate U-turn at an intermediate waypoint. At an intermediate waypoint, if the value of this property is false, each returned route may continue straight ahead or turn to either side but may not U-turn. This property has no effect if only two waypoints are specified.
   /// same as 'not continueStraight' on Android
   final bool? allowsUTurnAtWayPoints;
@@ -94,6 +100,9 @@ class MapBoxOptions {
       this.longPressDestinationEnabled,
       this.simulateRoute,
       this.isOptimized,
+      this.maxHeight,
+      this.maxLength,
+      this.maxWeight,
       this.mapStyleUrlDay,
       this.mapStyleUrlNight,
       this.enableFreeDriveMode,
@@ -138,6 +147,10 @@ class MapBoxOptions {
     if (this.simulateRoute != null)
       optionsMap['simulateRoute'] = this.simulateRoute;
     if (this.isOptimized != null) optionsMap['isOptimized'] = this.isOptimized;
+
+    if (maxHeight != null) optionsMap['maxHeight'] = maxHeight;
+    if (maxWeight != null) optionsMap['maxWeight'] = maxWeight;
+    if (maxLength != null) optionsMap['maxLength'] = maxLength;
 
     addIfNonNull('padding', <double?>[
       padding?.top,