+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.3
+
+* Swap length with width
+
+
+## 0.0.2
+
+* Add max height / length / weight arguments to query 
+
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* initial release.